Code updated

+++ admin/add_student.php
@@ -123,9 +123,8 @@ $result = $conn->query($query);
             </div>
         </form>
 
-        <h2 class="text-2xl text-center mt-10">All Students</h2>
         <div class="overflow-x-auto mt-4">
-            <table class="min-w-full bg-white">
+            <table id="studentsTable" class="min-w-full bg-white">
                 <thead>
                     <tr>
                         <th class="py-2 px-4 border">ID</th>
@@ -154,8 +153,9 @@ $result = $conn->query($query);
                                 <a href="edit_student.php?id=<?php echo $student['id']; ?>" class="text-blue-500 hover:underline">
                                     <i class="fas fa-edit"></i> Edit
                                 </a>
-                                <a href="add_student.php?delete_id=<?php echo $student['id']; ?>" class="text-red-500 hover:underline">
-                                    <i class="fas fa-trash"></i> Delete
+
+                                <a href="add_student.php?delete_id=<?php echo $student['id']; ?>" class="text-red-500 hover:underline" onclick="return confirm('Are you sure you want to delete this student?');">
+                                    <i class="fas fa-trash-alt"></i> Delete
                                 </a>
                             </td>
                         </tr>
